+++ src/Sidebar.css
@@ -1,0 +1,45 @@
+.Sidebar {
+    height: 100%;
+    width: 260px;
+    position: fixed;
+    border: 1px solid lightgray;
+    top: 50;
+    left: 0;
+    background-color: #ffffff;
+    overflow-x: hidden;
+    padding-top: 20px;
+  }
+  .SidebarList {
+    height: 15%;
+    padding: 0;
+    width: 100%;
+  }
+  
+  .SidebarList .SidebarRow {
+    width: 100%;
+    height: 60%;
+    list-style-type: none;
+    display: flex;
+    flex-direction: row;
+    margin: 0%;
+    justify-content: center;
+    align-items: center;
+    font-family: Verdana;
+  }
+  
+  .SidebarList .SidebarRow:hover {
+    cursor: pointer;
+    background-color: rgb(241, 241, 241);
+  }
+  
+  .SidebarList #active {
+    background-color: rgb(241, 241, 241);
+  }
+  .SidebarRow #RowIcon {
+    flex: 30%;
+    display: grid;
+    place-items: center;
+  }
+  .SidebarRow #RowTitle {
+    flex: 70%;
+  }
