+++ src/pages/Footer.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './Footer.css';
+function Footer() {
+    return (
+        <div className='footer'>
+         <p>© 2021 Recess  • All rights reserved  </p>   
+         <br></br>
+         <p>Privacy Policy  •   Terms & Conditions   •   About  •   Help • Contact </p>
+        </div>
+    )
+}
+
+export default Footer;
